% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integralBayes.R
\name{integralBayes}
\alias{integralBayes}
\title{integralBayes}
\usage{
integralBayes(meanval, sdval)
}
\arguments{
\item{meanval}{value of the mean}

\item{sdval}{value of the standard deviation}
}
\value{
a plot with appropriate shading with labels
}
\description{
Produces a plot with shading under the curve
}
\examples{
\dontrun{integralBayes(0.0, 0.2)}
}
