% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runningprop.R
\name{runningprop}
\alias{runningprop}
\title{runningprop}
\usage{
runningprop(N, pHeads)
}
\arguments{
\item{N}{number of flips}

\item{pHeads}{probability of heads}
}
\value{
a plot with running proportion of heads
}
\description{
Simulates coin flips
}
\examples{
\dontrun{runningprop(N = 100, pHeads = 0.5)}
}
